+++ module/spiff-element-units-rs/src/lib.rs
@@ -3,7 +3,7 @@ use std::iter::zip;
 
 mod cache;
 mod config;
-mod domain;
+mod basis;
 mod element_units;
 mod manifest;
 mod reader;
